Move Generalize tactic to SimpleDeducedSteps
