+++ lisa-sets2/src/main/scala/lisa/maths/Quantifiers.scala
@@ -309,55 +309,4 @@ object Quantifiers extends lisa.Main {
 
    */
 
-  /**
-   * Quantify all variables in a formula on the right side of the premise sequent.
-   *
-   * <pre>
-   *         Γ ⊢ φ, Δ
-   * -------------------------- x, y, ..., z do not appear in Γ
-   *  Γ ⊢ ∀x.∀y. ... ∀z. φ, Δ
-   * </pre>
-   */
-  object quantifyAll extends ProofFactSequentTactic:
-    def apply(using lib: Library, proof: lib.Proof)(premiseStep: proof.Fact)(conclusion: Sequent) =
-      def isQuantifiedOf(target: Expr[Prop], pivot: Expr[Prop], vars: List[Variable[Ind]] = Nil): Option[List[Variable[Ind]]] =
-        target match
-          case ∀(x, inner) =>
-            val next = x :: vars
-            if isSame(inner, pivot) then Some(next) else isQuantifiedOf(inner, pivot, next)
-          case _ => None
-      val premise = proof.getSequent(premiseStep)
-      val difference = premise.right -- conclusion.right
-
-      if difference.isEmpty then Restate(using lib, proof)(premiseStep)(conclusion)
-      else if difference.size > 1 then proof.InvalidProofTactic(s"There must be only one formula to quantify over between the premise and the conclusion. Found: \n${Printing.printList(difference)}")
-      else
-        val rdifference = conclusion.right -- premise.right
-        if rdifference.size != 1 then proof.InvalidProofTactic(s"There must be only one formula to quantify over between the premise and the conclusion. Found: \n${Printing.printList(rdifference)}")
-        else
-          val pivot = difference.head
-          val target = rdifference.head
-          val varsOption = isQuantifiedOf(target, pivot)
-
-          if varsOption.isEmpty then proof.InvalidProofTactic("Could not find a formula to quantify over in the conclusion.")
-          else
-            val vars = varsOption.get
-            val conflicts = vars.toSet `intersect` premise.left.flatMap(_.freeVars)
-
-            if conflicts.nonEmpty then proof.InvalidProofTactic(s"Variable(s) ${conflicts.mkString(", ")} to be quantified appear in the LHS of the conclusion.")
-            else
-              // safe, proceed
-              TacticSubproof:
-                val vars = varsOption.get
-                lib.have(premise) by Restate.from(premiseStep)
-
-                val base = premise ->> pivot
-
-                vars.foldLeft(pivot): (pivot, v) =>
-                  val quant = ∀(v, pivot)
-                  lib.thenHave(base +>> quant) by RightForall.withParameters(pivot, v)
-                  quant
-
-                lib.thenHave(conclusion) by Restate
-
 }
